pyproject.toml
[tool.poetry]
name = "py-aiger-abc"
readme="README.md"
version = "0.1.0"
description = "Bridge to using Berkeley-ABC with py-aiger."
repository = "https://github.com/mvcisback/py-aiger-abc"
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
license = "MIT"
packages = [
    { include = "aiger_abc" },
]

[tool.poetry.dependencies]
python = "^3.7"
py-aiger = "^3.5"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.7"
pytest-flake8 = "^1.0"
pytest-sugar = "^0.9.2"
pytest-xdist = "^1.29"
codecov = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
